Add files via upload
A portfolio project courtesy of Absent Data , demonstrating data analysis skills and knowledge of tools such as SQL data exploration and query, Power BI for analysis and visualization. Analyzing employee data regarding their health and absenteeism patterns with the view of awarding an annual incentive bonus to staff that meet certain health and low absenteeism criteria. The aim of this query is to identify and extract insightful data that would inform decision making by HR as to what constitutes 'healthy' so as to identify staff to be given this bonus.